24 Day 7 Part 1 Complete
Not First try but it was simple fix.

I needed to upgrade the answer int to a long but when I did I got to adjust the testAnswer in the function to long too so first attempt was doing weird out of bounds math with int number not going high enough.

+++ 2024/AOC2407.cs
@@ -81,7 +81,7 @@ namespace CodeTAF
         bool checkForValidOps(string curOps, (int depth, long answer, int[] equationNums) target ) {
 
             if (curOps.Length == target.depth) {
-                int answer = (curOps[0] == '*')? (target.equationNums[0] * target.equationNums[1]) : (target.equationNums[0] + target.equationNums[1]);
+                long answer = (curOps[0] == '*')? (target.equationNums[0] * target.equationNums[1]) : (target.equationNums[0] + target.equationNums[1]);
 
                 //check for valid solution
                 for (int i = 1; i < curOps.Length; i++) {
